To-Control-Win-Server/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mouse==0.7.1
mss==6.1.0
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
playsound==1.3.0
pytesseract==0.3.8
waitress==2.0.0
Werkzeug==2.0.1
